crypto: Move some olm-related tests to a separate file
